requirements.txt
numpy==1.26.0
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
googletrans==3.1.0a0
easyocr==1.7.1

torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0